requirements.txt
pipecat-ai[silero]
openai
fastapi
uvicorn[standard]
websockets
python-multipart
aiofiles
requests
sounddevice
numpy
python-dotenv
loguru